Allow dotenv config to overwrite env config
This is necessary in the current dev container setup for a slightly
stupid reason:

In the Docker Compose setup, the DB URL is determined by the name chosen
for the DB service. Therefore this information is derived from the
compose file itself, which means that the logical way to pass the config
into the backend is to set an env var for the backend service, again in
the compose file.

In the dev container, we are running a derivative of the backend
service. Therefore the environment is inherited from the Compose file.
Therefore the DB URL setting from the compose file is present in the
environment of the dev container. Now say we want to run tests in a
different container, and set a different DB_URL in .env.test -- it
doesn't work, because the Vapor DotEnv provider won't overwrite the
value from the environment.

That's why this change is necessary. It is unfortunate though because
now the config is spread in sort of surprising ways across multiple
places. This needs more thinking I think.